+++ CHANGELOG.md
@@ -11,3 +11,10 @@
 - versioning 
 
 ## 2021-03-25 [0.1.2]
+
+## 2022-02-04 [0.1.3]  
+- Add the attribute JobTitle
+- Add the cronjob to check if users are still active
+- Adjust the Client(local code) to the changes made to the API 
+- Develop a cronjob plugin and connect it with this plugin(see CronJob plugin)
+- Fix Exception handling